+++ packages/insomnia/src/ui/index.tsx
@@ -40,7 +40,7 @@ import { ErrorRoute } from './routes/error';
 import Root from './routes/root';
 import { initializeSentry } from './sentry';
 
-console.log(routes);
+console.log({ routes });
 
 initializeSentry();
 initializeLogging();
@@ -56,7 +56,6 @@ if (prevLocationHistoryEntry && matchPath({ path: '/organization/:organizationId
 }
 
 const router = createMemoryRouter(
-  // @TODO - Investigate file based routing to generate these routes:
   [
     {
       path: '/',
@@ -64,219 +63,6 @@ const router = createMemoryRouter(
       element: <Root />,
       errorElement: <ErrorRoute />,
       children: routes,
-      // children: [
-      //   {
-      //     path: 'organization',
-      //     children: [
-      //       {
-      //         path: ':organizationId',
-      //         children: [
-      //           {
-      //             index: true,
-      //             loader: async (...args) => (await import('./routes/project')).indexLoader(...args),
-      //           },
-      //           {
-      //             path: 'project',
-      //             children: [
-      //               {
-      //                 path: ':projectId',
-      //                 id: '/project/:projectId',
-      //                 loader: async (...args) =>
-      //                   (await import('./routes/project')).loader(...args),
-      //                 element: (
-      //                   <Suspense fallback={<AppLoadingIndicator />}>
-      //                     <Project />
-      //                   </Suspense>
-      //                 ),
-      //                 children: [
-      //                   {
-      //                     path: 'delete',
-      //                     action: async (...args) =>
-      //                       (await import('./routes/actions')).deleteProjectAction(
-      //                         ...args
-      //                       ),
-      //                   },
-      //                   {
-      //                     path: 'rename',
-      //                     action: async (...args) =>
-      //                       (await import('./routes/actions')).renameProjectAction(
-      //                         ...args
-      //                       ),
-      //                   },
-      //                 ],
-      //               },
-      //               {
-      //                 path: ':projectId/workspace',
-      //                 children: [
-      //                   {
-      //                     path: ':workspaceId',
-      //                     loader: async (...args) => (await import('./routes/workspace')).workspaceLoader(...args),
-      //                     children: [
-      //                       {
-      //                         path: `${ACTIVITY_DEBUG}`,
-      //                         element: (
-      //                           <Suspense fallback={<AppLoadingIndicator />}>
-      //                             <Debug />
-      //                           </Suspense>
-      //                         ),
-      //                       },
-      //                       {
-      //                         path: `${ACTIVITY_SPEC}`,
-      //                         element: (
-      //                           <Suspense fallback={<AppLoadingIndicator />}>
-      //                             <Design />
-      //                           </Suspense>
-      //                         ),
-      //                       },
-      //                       {
-      //                         path: 'test/*',
-      //                         loader: async (...args) =>  (await import('./routes/unit-test')).loader(...args),
-      //                         element: (
-      //                           <Suspense fallback={<AppLoadingIndicator />}>
-      //                             <UnitTest />
-      //                           </Suspense>
-      //                         ),
-      //                         children: [
-      //                           {
-      //                             index: true,
-      //                             loader: async (...args) => (await import('./routes/test-suite')).indexLoader(...args),
-      //                           },
-      //                           {
-      //                             path: 'test-suite',
-      //                             children: [
-      //                               {
-      //                                 index: true,
-      //                                 loader: async (...args) => (await import('./routes/test-suite')).indexLoader(...args),
-      //                               },
-      //                               {
-      //                                 path: 'new',
-      //                                 action: async (...args) => (await import('./routes/actions')).createNewTestSuiteAction(...args),
-      //                               },
-      //                               {
-      //                                 path: ':testSuiteId',
-      //                                 id: ':testSuiteId',
-      //                                 loader: async (...args) => (await import('./routes/test-suite')).loader(...args),
-      //                                 children: [
-      //                                   {
-      //                                     index: true,
-      //                                     loader: async (...args) => (await import('./routes/test-results')).indexLoader(...args),
-      //                                   },
-      //                                   {
-      //                                     path: 'test-result',
-      //                                     children: [
-      //                                       {
-      //                                         path: ':testResultId',
-      //                                         id: ':testResultId',
-      //                                         loader: async (...args) => (await import('./routes/test-results')).loader(...args),
-      //                                       },
-      //                                     ],
-      //                                   },
-      //                                   {
-      //                                     path: 'delete',
-      //                                     action: async (...args) => (await import('./routes/actions')).deleteTestSuiteAction(...args),
-      //                                   },
-      //                                   {
-      //                                     path: 'rename',
-      //                                     action: async (...args) => (await import('./routes/actions')).renameTestSuiteAction(...args),
-      //                                   },
-      //                                   {
-      //                                     path: 'run-all-tests',
-      //                                     action: async (...args) => (await import('./routes/actions')).runAllTestsAction(...args),
-      //                                   },
-      //                                   {
-      //                                     path: 'test',
-      //                                     children: [
-      //                                       {
-      //                                         path: 'new',
-      //                                         action: async (...args) => (await import('./routes/actions')).createNewTestAction(...args),
-      //                                       },
-      //                                       {
-      //                                         path: ':testId',
-      //                                         children: [
-      //                                           {
-      //                                             path: 'delete',
-      //                                             action: async (...args) => (await import('./routes/actions')).deleteTestAction(...args),
-      //                                           },
-      //                                           {
-      //                                             path: 'update',
-      //                                             action: async (...args) => (await import('./routes/actions')).updateTestAction(...args),
-      //                                           },
-      //                                           {
-      //                                             path: 'run',
-      //                                             action: async (...args) => (await import('./routes/actions')).runTestAction(...args),
-      //                                           },
-      //                                         ],
-      //                                       },
-      //                                     ],
-      //                                   },
-      //                                 ],
-      //                               },
-      //                             ],
-      //                           },
-      //                         ],
-      //                       },
-      //                       {
-      //                         path: 'duplicate',
-      //                         action: async (...args) =>
-      //                           (
-      //                             await import('./routes/actions')
-      //                           ).duplicateWorkspaceAction(...args),
-      //                       },
-      //                     ],
-      //                   },
-      //                   {
-      //                     path: 'new',
-      //                     action: async (...args) =>
-      //                       (await import('./routes/actions')).createNewWorkspaceAction(
-      //                         ...args
-      //                       ),
-      //                   },
-      //                   {
-      //                     path: 'delete',
-      //                     action: async (...args) =>
-      //                       (await import('./routes/actions')).deleteWorkspaceAction(
-      //                         ...args
-      //                       ),
-      //                   },
-      //                   {
-      //                     path: 'update',
-      //                     action: async (...args) =>
-      //                       (await import('./routes/actions')).updateWorkspaceAction(
-      //                         ...args
-      //                       ),
-      //                   },
-      //                 ],
-      //               },
-      //               {
-      //                 path: 'new',
-      //                 action: async (...args) =>
-      //                   (await import('./routes/actions')).createNewProjectAction(
-      //                     ...args
-      //                   ),
-      //               },
-      //               {
-      //                 path: ':projectId/remote-collections',
-      //                 loader: async (...args) =>
-      //                   (
-      //                     await import('./routes/remote_collections')
-      //                   ).remoteCollectionsLoader(...args),
-      //                 children: [
-      //                   {
-      //                     path: 'pull',
-      //                     action: async (...args) =>
-      //                       (
-      //                         await import('./routes/remote_collections')
-      //                       ).pullRemoteCollectionAction(...args),
-      //                   },
-      //                 ],
-      //               },
-      //             ],
-      //           },
-      //         ],
-      //       },
-      //     ],
-      //   },
-      // ],
     },
   ],
   {
