+++ packages/insomnia/src/ui/routes/organization/$organizationId/project/$projectId/delete.tsx
@@ -5,7 +5,7 @@ import { SegmentEvent, trackSegmentEvent } from '@insomnia/ui/analytics';
 import { invariant } from '@remix-run/router';
 import { ActionFunction, redirect } from 'react-router-dom';
 
-export const deleteProjectAction: ActionFunction = async ({ params }) => {
+export const action: ActionFunction = async ({ params }) => {
   const { organizationId, projectId } = params;
   invariant(organizationId, 'Organization ID is required');
   invariant(projectId, 'Project ID is required');
